# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Course & Training Management System API (ИПР №1)",
    "version": "1.0.0",
    "description": "Безопасный REST API для системы управления курсами, личным кабинетом студента, кабинетом преподавателя и динамическим RBAC (Role-Based Access Control). Спецификация сгенерирована из декларативных аннотаций Swaggo (// Method Godoc, @Summary, @Description, @Tags, @Param, @Success, @Failure, @Router) в контроллерах."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1 Base Endpoint"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Введите JWT Access токен (Bearer <token>)"
      }
    },
    "schemas": {
...
      "CreateRoleRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "moderator"
          },
          "description": {
            "type": "string",
            "example": "Модератор обучающих курсов и заявок студентов"
          }
        },
        "example": {
          "name": "moderator",
          "description": "Модератор обучающих курсов и заявок студентов"
        }
      },
      "UpdateRoleRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "content_manager"
          },
          "description": {
            "type": "string",
            "example": "Управление публикацией учебных программ"
          }
        },
        "example": {
          "name": "content_manager",
          "description": "Управление публикацией учебных программ"
        }
      },
...
      "AssignRoleRequest": {
        "type": "object",
        "required": [
          "role_id"
        ],
        "properties": {
          "role_id": {
            "type": "string",
            "example": "role-uuid-instructor"
          }
        },
        "example": {
          "role_id": "role-uuid-instructor"
        }
      },
...
      "RoleResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "role-uuid-instructor"
              },
              "name": {
                "type": "string",
                "example": "instructor"
              },
              "description": {
                "type": "string",
                "example": "Преподаватель курсов"
              },
              "permissions": {
                "type": "array",
                "items": {
                  "type": "object"
                },
                "example": [
                  {
                    "id": "perm-1",
                    "code": "courses:create",
                    "description": "Создание курсов"
                  },
                  {
                    "id": "perm-2",
                    "code": "courses:edit",
                    "description": "Редактирование своих курсов"
                  }
                ]
              }
            }
          }
        },
        "example": {
          "success": true,
          "data": {
            "id": "role-uuid-instructor",
            "name": "instructor",
            "description": "Преподаватель курсов",
            "permissions": [
              {
                "id": "perm-1",
                "code": "courses:create",
                "description": "Создание курсов"
              },
              {
                "id": "perm-2",
                "code": "courses:edit",
                "description": "Редактирование своих курсов"
              },
              {
                "id": "perm-3",
                "code": "courses:view_students",
                "description": "Просмотр студентов курса"
              }
            ]
          }
        }
      },
      "RoleListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleResponse"
            }
          }
        },
        "example": {
          "success": true,
          "data": [
            {
              "id": "role-1",
              "name": "admin",
              "description": "Полный администратор платформы"
            },
            {
              "id": "role-2",
              "name": "instructor",
              "description": "Преподаватель курсов"
            },
            {
              "id": "role-3",
              "name": "student",
              "description": "Студент / Слушатель"
            }
          ]
        }
      },
      "AssignRoleResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Пользователю успешно назначена роль instructor."
          },
          "data": {
            "type": "object",
            "properties": {
              "user_id": {
                "type": "string",
                "example": "usr-uuid-1"
              },
              "email": {
                "type": "string",
                "example": "student@example.com"
              },
              "role": {
                "type": "string",
                "example": "instructor"
              }
            }
          }
        },
        "example": {
          "success": true,
          "message": "Пользователю успешно назначена роль instructor.",
          "data": {
            "user_id": "usr-uuid-1",
            "email": "student@example.com",
            "role": "instructor"
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Операция успешно выполнена."
          }
        },
        "example": {
          "success": true,
          "message": "Операция успешно выполнена."
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Неверные параметры запроса."
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        },
        "example": {
          "success": false,
          "error": {
            "code": "VALIDATION_ERROR",
            "message": "Неверные параметры запроса."
          }
        }
      }
    }
  },
  "paths": {
...
    "/roles": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "Получить список всех ролей",
        "description": "Возвращает роли с привязанными правами (пермишн roles:manage)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "models.RoleListResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListResponse"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "role-1",
                      "name": "admin",
                      "description": "Полный администратор платформы"
                    },
                    {
                      "id": "role-2",
                      "name": "instructor",
                      "description": "Преподаватель курсов"
                    },
                    {
                      "id": "role-3",
                      "name": "student",
                      "description": "Студент / Слушатель"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "roles"
        ],
        "summary": "Создать новую роль",
        "description": "Создает новую системную или пользовательскую роль (пермишн roles:manage)",
        "parameters": [],
        "responses": {
          "201": {
            "description": "models.RoleResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "role-uuid-instructor",
                    "name": "instructor",
                    "description": "Преподаватель курсов",
                    "permissions": [
                      {
                        "id": "perm-1",
                        "code": "courses:create",
                        "description": "Создание курсов"
                      },
                      {
                        "id": "perm-2",
                        "code": "courses:edit",
                        "description": "Редактирование своих курсов"
                      },
                      {
                        "id": "perm-3",
                        "code": "courses:view_students",
                        "description": "Просмотр студентов курса"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          },
          "409": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "Данные роли",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleRequest"
              },
              "example": {
                "name": "moderator",
                "description": "Модератор обучающих курсов и заявок студентов"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/roles/{id}": {
      "put": {
        "tags": [
          "roles"
        ],
        "summary": "Обновить роль",
        "description": "Обновляет название или описание роли (запрещено переименовывать базовые роли)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role ID",
            "schema": {
              "type": "string",
              "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
            },
            "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
          }
        ],
        "responses": {
          "200": {
            "description": "models.RoleResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "role-uuid-instructor",
                    "name": "instructor",
                    "description": "Преподаватель курсов",
                    "permissions": [
                      {
                        "id": "perm-1",
                        "code": "courses:create",
                        "description": "Создание курсов"
                      },
                      {
                        "id": "perm-2",
                        "code": "courses:edit",
                        "description": "Редактирование своих курсов"
                      },
                      {
                        "id": "perm-3",
                        "code": "courses:view_students",
                        "description": "Просмотр студентов курса"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          },
          "404": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "Новые данные роли",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleRequest"
              },
              "example": {
                "name": "content_manager",
                "description": "Управление публикацией учебных программ"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "roles"
        ],
        "summary": "Удалить роль",
        "description": "Удаляет кастомную роль (запрещено удалять системные роли и роли, назначенные пользователям)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role ID",
            "schema": {
              "type": "string",
              "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
            },
            "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
          }
        ],
        "responses": {
          "200": {
            "description": "models.SuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "message": "Операция успешно выполнена."
                }
              }
            }
          },
          "400": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          },
          "404": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/permissions": {
      "get": {
        "tags": [
          "roles"
        ],
        "summary": "Получить список всех прав",
        "description": "Возвращает справочник всех атомарных прав системы (пермишн permissions:view)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "models.PermissionsListResponse",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/roles/{id}/permissions": {
      "post": {
        "tags": [
          "roles"
        ],
        "summary": "Назначить права роли",
        "description": "Динамическое назначение атомарных прав роли (пермишн roles:manage)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role ID",
            "schema": {
              "type": "string",
              "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
            },
            "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
          }
        ],
        "responses": {
          "200": {
            "description": "models.RoleResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "role-uuid-instructor",
                    "name": "instructor",
                    "description": "Преподаватель курсов",
                    "permissions": [
                      {
                        "id": "perm-1",
                        "code": "courses:create",
                        "description": "Создание курсов"
                      },
                      {
                        "id": "perm-2",
                        "code": "courses:edit",
                        "description": "Редактирование своих курсов"
                      },
                      {
                        "id": "perm-3",
                        "code": "courses:view_students",
                        "description": "Просмотр студентов курса"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "Список прав",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/users/{id}/role": {
      "patch": {
        "tags": [
          "roles"
        ],
        "summary": "Назначить роль пользователю",
        "description": "Изменение роли пользователя (пермишн users:manage_roles)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "User ID",
            "schema": {
              "type": "string",
              "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
            },
            "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
          }
        ],
        "responses": {
          "200": {
            "description": "models.AssignRoleResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignRoleResponse"
                },
                "example": {
                  "success": true,
                  "message": "Пользователю успешно назначена роль instructor.",
                  "data": {
                    "user_id": "usr-uuid-1",
                    "email": "student@example.com",
                    "role": "instructor"
                  }
                }
              }
            }
          },
          "404": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "Роль",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleRequest"
              },
              "example": {
                "role_id": "role-uuid-instructor"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}